// src/components/BaseTable/models/AdvancedTableSettings.ts
export default interface AdvancedTableSettings {
  marginTop?: string;
  noBorder?: boolean;
  pinColumns?: boolean;
  alignCenterInLine?: boolean;
  currentSortId?: string;
  contrastRow?: boolean;
  indexUseOriginalOrder?: boolean;
  enableRowDragDrop?: boolean;
  currentUsername?: string;
  showIndex?: boolean;
  selectNewOptionOnAdd?: boolean;
  canAddListOptions?: boolean;
  showFooter?: boolean;
  rowIdProperty?: string;
  tableId?: string;
  focusedRowIndex?: number;
}